// src/types.ts
export enum Field {
  THEME = "Theme",
  SUBTHEME = "Sub-Theme",
  REASON = "Reason",
  LANGUAGE = "Language",
  SOURCE = "Source",
  RATING = "Rating",
  TIME_PERIOD = "Time Period",
  CUSTOMERID = "Customer ID",
}

export enum Condition {
  EQUALS = "Equals",
  DOES_NOT_EQUAL = "Does not equal",
  LIKE = "Like",
  NOT_LIKE = "Not like",
  IS_EMPTY = "Is Empty",
  IS = "Is",
  IS_NOT = "Is not",
};